Move set_random_seed to utils
